Add files via upload
Here is my center module for the turreted robot! it uses a 4 bar mechanism, with a dead axle pivot. the dead axle is belted to the end effector  to keep it straight at the ground at all times. i will probably add a wrist servo to the arm anways but this makes it easier to program and looks cool